sql uplodaed

+++ appcrud/src/pages/ViewUser.js
@@ -2,6 +2,7 @@ import React, {useState, useEffect} from 'react';
 import {Link} from 'react-router-dom';
 import axios from 'axios';
 import swal from 'sweetalert';
+import Swal from 'sweetalert2'
 
 function ViewUser() {
 
@@ -22,21 +23,33 @@ function ViewUser() {
 
     const deleteUser = (e, id) => {
         e.preventDefault();
-        
         const thisClicked = e.currentTarget;
-        thisClicked.innerText = "Deleting";
+        
+        Swal.fire({
+          title: "Are you sure?",
+          text: "You won't be able to Delete this!",
+          icon: "warning",
+          showCancelButton: true,
+          confirmButtonColor: "#3085d6",
+          cancelButtonColor: "#d33",
+          confirmButtonText: "Yes, delete it!"
+        }).then((result) => {
+          if (result.isConfirmed) {
+            thisClicked.innerText = "Deleting";
 
-        axios.delete(`/api/delete-user/${id}`).then(res=>{
-            if(res.data.status === 200)
-            {
-                swal("Deleted!",res.data.message,"success");
-                thisClicked.closest("tr").remove();
-            }
-            else if(res.data.status === 404)
-            {
-                swal("Error",res.data.message,"error");
-                thisClicked.innerText = "Delete";
-            }
+            axios.delete(`/api/delete-user/${id}`).then(res=>{
+                if(res.data.status === 200)
+                {
+                    swal("Deleted!",res.data.message,"success");
+                    thisClicked.closest("tr").remove();
+                }
+                else if(res.data.status === 404)
+                {
+                    swal("Error",res.data.message,"error");
+                    thisClicked.innerText = "Delete";
+                }
+            });
+          }
         });
     }
